refactor: ドメイン層と共有ユーティリティ層への責務分離とファイル移動
