+++ Frontend/src/domain/models/layoutTypes.ts
