+++ Frontend/src/domain/models/terms.ts
