+++ Frontend/src/shared/utils/verifySimilarity.ts
